Convert shouldWriteNoRecords test into unit test
